refactor: refactor the code for supporting more nested list types
