+++ lib/nested_choice_list.dart
@@ -3,6 +3,7 @@ library nested_choice_list;
 export 'src/navigation_path/navigation_path.dart';
 export 'src/nested_choice_entity.dart';
 export 'src/nested_choice_list.dart';
+export 'src/nested_choice_list_type.dart';
 export 'src/nested_list_style/navigation_path_item_style.dart';
 export 'src/nested_list_style/nested_list_item_style.dart';
 export 'src/nested_list_style/nested_list_searchfield_style.dart';
