+++ lib/src/nested_choice_list_type.dart
@@ -1,0 +1,10 @@
+/// An enumeration that defines the types of nested choice lists.
+///
+/// The `NestedChoiceListType` enum provides two options:
+///
+/// - `navigation`: Represents a nested choice list that uses navigation.
+/// - `expandable`: Represents a nested choice list that can be expanded.
+enum NestedChoiceListType {
+  navigation,
+  expandable,
+}
