+++ test/src/nested_navigation_choice_list/nested_navigation_choice_list_test.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:nested_choice_list/src/nested_choice_entity.dart';
-import 'package:nested_choice_list/src/nested_list_view.dart';
+import 'package:nested_choice_list/src/nested_navigation_choice_list/nested_navigation_choice_list.dart';
 import 'package:nested_choice_list/src/search_field/searchfield_position.dart';
 
 void main() {
-  group('NestedListView Tests', () {
+  group('NestedNavigationChoiceList Tests', () {
     testWidgets('renders correctly with default parameters', (tester) async {
       await tester.pumpWidget(
         const MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: [NestedChoiceEntity(value: '1', label: 'Item 1')],
           ),
         ),
@@ -21,7 +21,7 @@ void main() {
     testWidgets('search functionality filters items correctly', (tester) async {
       await tester.pumpWidget(
         const MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: [
               NestedChoiceEntity(value: '1', label: 'Item 1'),
               NestedChoiceEntity(value: '2', label: 'Item 2'),
@@ -35,8 +35,6 @@ void main() {
       await tester.enterText(find.byType(TextField), 'Item 1');
       await tester.pumpAndSettle();
 
-      // finds two items with the text 'Item 1' because the search query
-      // is 'Item 1'
       expect(find.text('Item 1'), findsNWidgets(2));
       expect(find.text('Item 2'), findsNothing);
     });
@@ -45,14 +43,14 @@ void main() {
       bool isSelectedAll = false;
       await tester.pumpWidget(
         MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: const [
               NestedChoiceEntity(value: '1', label: 'Item 1'),
               NestedChoiceEntity(value: '2', label: 'Item 2'),
             ],
             enableMultiSelect: true,
-            onSelectAllCallback: ({required isSelected, required items}) {
-              isSelectedAll = isSelected;
+            onSelectionChange: (items) {
+              isSelectedAll = items.length == 2;
             },
           ),
         ),
@@ -68,9 +66,9 @@ void main() {
       NestedChoiceEntity? tappedItem;
       await tester.pumpWidget(
         MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: const [NestedChoiceEntity(value: '1', label: 'Item 1')],
-            onTapItem: (item, context) {
+            onTapItem: (item) {
               tappedItem = item;
             },
           ),
@@ -84,16 +82,15 @@ void main() {
     });
 
     testWidgets('item toggle selection callback is triggered', (tester) async {
-      NestedChoiceEntity? toggledItem;
+      List<NestedChoiceEntity> selectedItems = [];
       await tester.pumpWidget(
         MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: const [NestedChoiceEntity(value: '1', label: 'Item 1')],
             enableMultiSelect: true,
-            onToggleSelection: (item) {
-              toggledItem = item;
+            onSelectionChange: (items) {
+              selectedItems = items;
             },
-            onSelectAllCallback: ({required isSelected, required items}) {},
           ),
         ),
       );
@@ -101,13 +98,14 @@ void main() {
       await tester.tap(find.text('Item 1'));
       await tester.pump();
 
-      expect(toggledItem?.label, 'Item 1');
+      expect(selectedItems.length, 1);
+      expect(selectedItems.first.label, 'Item 1');
     });
 
     testWidgets('leading widget builder works correctly', (tester) async {
       await tester.pumpWidget(
         MaterialApp(
-          home: NestedListView(
+          home: NestedNavigationChoiceList(
             items: const [NestedChoiceEntity(value: '1', label: 'Item 1')],
             itemLeadingBuilder: (context, item, index) {
               return const Icon(Icons.star);
@@ -117,6 +115,77 @@ void main() {
       );
 
       expect(find.byIcon(Icons.star), findsOneWidget);
+    });
+
+    testWidgets('navigation path is displayed correctly', (tester) async {
+      await tester.pumpWidget(
+        const MaterialApp(
+          home: NestedNavigationChoiceList(
+            items: [
+              NestedChoiceEntity(
+                value: '1',
+                label: 'Item 1',
+                children: [
+                  NestedChoiceEntity(value: '1.1', label: 'Item 1.1'),
+                ],
+              ),
+            ],
+            showNavigationPath: true,
+          ),
+        ),
+      );
+
+      await tester.tap(find.text('Item 1'));
+      await tester.pumpAndSettle();
+
+      expect(find.text('Item 1'), findsOneWidget);
+      expect(find.text('Item 1.1'), findsOneWidget);
+    });
+
+    testWidgets('onNavigationChange callback is triggered', (tester) async {
+      int navigationPageIndex = 0;
+      await tester.pumpWidget(
+        MaterialApp(
+          home: NestedNavigationChoiceList(
+            items: const [
+              NestedChoiceEntity(
+                value: '1',
+                label: 'Item 1',
+                children: [
+                  NestedChoiceEntity(value: '1.1', label: 'Item 1.1'),
+                ],
+              ),
+            ],
+            onNavigationChange: (pageIndex) {
+              navigationPageIndex = pageIndex;
+            },
+          ),
+        ),
+      );
+
+      await tester.tap(find.text('Item 1'));
+      await tester.pumpAndSettle();
+
+      expect(navigationPageIndex, 1);
+    });
+
+    testWidgets('selected items chip list is displayed correctly',
+        (tester) async {
+      await tester.pumpWidget(
+        const MaterialApp(
+          home: NestedNavigationChoiceList(
+            items: [
+              NestedChoiceEntity(value: '1', label: 'Item 1'),
+              NestedChoiceEntity(value: '2', label: 'Item 2'),
+            ],
+            enableMultiSelect: true,
+            selectedItems: [NestedChoiceEntity(value: '1', label: 'Item 1')],
+          ),
+        ),
+      );
+
+      expect(find.text('Item 1'), findsNWidgets(2));
+      expect(find.text('Item 2'), findsOneWidget);
     });
   });
 }
